social_media_downloader/requirements.txt
# 社交媒体统一下载工具 - 依赖列表

# 核心下载引擎
yt-dlp>=2026.3.0
instaloader>=4.15.0

# 中国平台
aiohttp>=3.9.0
aiofiles>=23.2.1
httpx[http2,socks]>=0.28.0
gmssl>=3.2.2

# B站
bilibili-api-python>=17.0.0

# 金融数据
akshare>=1.18.0

# 通用
requests>=2.31.0
lxml>=5.0.0
pyyaml>=6.0.0
rich>=13.7.0
click>=8.0.0

# 可选依赖
pyperclip>=1.8.0  # 剪贴板支持
tqdm>=4.60.0      # 进度条
